Started working on the adventure portion.
- Some loocations can have an adventure.
- Players an embark on an adventure.
- Canceling needs to be re-worked such that we store what adventure is
  in progress. we could technically set the in_progress to false.
- Character xp, skills, gold rushes and drops all now take inton account
  the adventure bonuses.
- we need a way store that data as part of the report instead of showing
  it in the chat.
- Moving some of the classes and code around to seperate concerns.
- Some tests are failing.

+++ database/migrations/2019_11_24_201001_create_monsters.php
@@ -25,7 +25,7 @@ class CreateMonsters extends Migration
             $table->Integer('max_level')->nullable()->default(0);
             $table->string('damage_stat');
             $table->integer('xp');
-            $table->integer('drop_check');
+            $table->float('drop_check');
             $table->integer('gold')->nullable()->default(10);
             $table->string('health_range');
             $table->string('attack_range');
